+++ lib/omni_video_player.dart
@@ -1,7 +1,6 @@
 export 'omni_video_player/controllers/global_playback_controller.dart';
 export 'omni_video_player/controllers/omni_playback_controller.dart';
 export 'omni_video_player/models/custom_player_widgets.dart';
-export 'omni_video_player/models/global_playback_control_settings.dart';
 export 'omni_video_player/models/omni_video_quality.dart';
 export 'omni_video_player/models/player_ui_visibility_options.dart';
 export 'omni_video_player/models/video_player_callbacks.dart';