+++ lib/src/widgets/player/video_player_error_placeholder.dart
@@ -1,16 +1,10 @@
-import 'package:android_intent_plus/android_intent.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:omni_video_player/omni_video_player/theme/omni_video_player_theme.dart';
-import 'package:omni_video_player/src/utils/logger.dart';
-import 'package:omni_video_player/src/video_player_initializer/video_player_initializer.dart';
-import 'package:url_launcher/url_launcher.dart';
 
 /// A default error placeholder widget for the video player.
 ///
 /// [VideoPlayerErrorPlaceholder] is shown when video playback fails due to an error.
-/// It displays a themed icon, error message, and optionally provides a button
-/// to open the video URL in an external video player app.
+/// It displays a themed icon, error message,
 ///
 /// The appearance and text are styled based on the current [OmniVideoPlayerTheme].
 class VideoPlayerErrorPlaceholder extends StatelessWidget {
@@ -20,17 +14,7 @@ class VideoPlayerErrorPlaceholder extends StatelessWidget {
   /// that allows the user to open the video in an external app.
   const VideoPlayerErrorPlaceholder({
     super.key,
-    required this.playerGlobalKey,
-    required this.videoUrlToOpenExternally,
-    required this.showRefreshButton,
   });
-
-  /// The video URL to open in an external player, if applicable.
-  final String? videoUrlToOpenExternally;
-
-  final bool showRefreshButton;
-
-  final GlobalKey<VideoPlayerInitializerState> playerGlobalKey;
 
   @override
   Widget build(BuildContext context) {
@@ -51,42 +35,6 @@ class VideoPlayerErrorPlaceholder extends StatelessWidget {
               style: TextStyle(color: theme.colors.textError),
               textAlign: TextAlign.center,
             ),
-            if (videoUrlToOpenExternally != null)
-              Padding(
-                padding: const EdgeInsets.only(top: 8.0),
-                child: ElevatedButton(
-                  onPressed: () async {
-                    if (defaultTargetPlatform == TargetPlatform.android) {
-                      final AndroidIntent intent = AndroidIntent(
-                        action: 'action_view',
-                        data: videoUrlToOpenExternally!,
-                      );
-                      await intent.launch();
-                    } else {
-                      try {
-                        await launchUrl(Uri.parse(videoUrlToOpenExternally!));
-                      } catch (e, s) {
-                        logger.w(
-                          'Failed to launch external video URL',
-                          error: e,
-                          stackTrace: s,
-                        );
-                      }
-                    }
-                  },
-                  child: Text(theme.labels.openExternalLabel),
-                ),
-              ),
-            if (showRefreshButton)
-              Padding(
-                padding: const EdgeInsets.only(top: 8.0),
-                child: OutlinedButton(
-                  onPressed: () async {
-                    playerGlobalKey.currentState?.refresh();
-                  },
-                  child: Text(theme.labels.refreshLabel),
-                ),
-              ),
           ],
         ),
       ),